move extensions to shared package
