+++ site/lib/src/extensions/registry.dart
@@ -3,13 +3,13 @@
 // found in the LICENSE file.
 
 import 'package:jaspr_content/jaspr_content.dart';
+import 'package:site_shared/extensions/attribute_processor.dart';
 import 'package:site_shared/extensions/code_block_processor.dart';
+import 'package:site_shared/extensions/header_extractor.dart';
+import 'package:site_shared/extensions/header_processor.dart';
+import 'package:site_shared/extensions/table_processor.dart';
 
-import 'attribute_processor.dart';
 import 'glossary_link_processor.dart';
-import 'header_extractor.dart';
-import 'header_processor.dart';
-import 'table_processor.dart';
 import 'tutorial_prefetch_processor.dart';
 import 'tutorial_structure_processor.dart';
